+++ src/ultrasphere/__init__.py
@@ -9,15 +9,23 @@ from ._creation import (
     standard,
     standard_prime,
 )
+from ._draw import draw
+from ._integral import integrate, roots
+from ._random import random_ball, random_sphere
 
 __all__ = [
     "SphericalCoordinates",
     "c_spherical",
+    "draw",
     "from_branching_types",
     "hopf",
+    "integrate",
     "polar",
     "random",
     "random",
+    "random_ball",
+    "random_sphere",
+    "roots",
     "standard",
     "standard_prime",
 ]
